storage: move get_old_value to MvccReader
Which lets us make write_cursor private.
