+++ src/storage/txn/actions/mod.rs
@@ -11,6 +11,5 @@ pub mod check_txn_status;
 pub mod cleanup;
 pub mod commit;
 pub mod gc;
-pub(crate) mod get_old_value;
 pub mod prewrite;
 pub mod tests;
